Remove branding images
